Update fetchSiteUiConfig - #4886
