+++ ui/app/src/services/configuration.ts
@@ -167,147 +167,31 @@ const messages = defineMessages({
   }
 });
 
-export function fetchSiteUiConfig(site: string): Observable<Pick<GlobalState['uiConfig'], 'preview' | 'launcher'>> {
-  return fetchConfigurationDOM(site, '/ui.xml', 'studio').pipe(
-    map((xml) => {
-      if (xml) {
-        const config = {
-          preview: {
-            toolbar: {
-              leftSection: null,
-              middleSection: null,
-              rightSection: null
-            },
-            toolsPanel: {
-              widgets: [
-                {
-                  id: 'craftercms.component.EmptyState',
-                  uiKey: -1,
-                  configuration: {
-                    title: messages.emptyUiConfigMessageTitle,
-                    subtitle: messages.emptyUiConfigMessageSubtitle
-                  }
-                }
-              ]
-            },
-            pageBuilderPanel: {
-              widgets: [
-                {
-                  id: 'craftercms.component.EmptyState',
-                  uiKey: -1,
-                  configuration: {
-                    title: messages.emptyUiConfigMessageTitle,
-                    subtitle: messages.emptyUiConfigMessageSubtitle
-                  }
-                }
-              ]
-            }
-          },
-          launcher: null,
-          dashboard: null,
-          references: {}
-        };
-        const arrays = ['widgets', 'roles', 'excludes', 'devices', 'values', 'siteCardMenuLinks', 'tools'];
-        const renameTable = { permittedRoles: 'roles' };
-        // Reading references
-        const references = {};
-        xml.querySelectorAll(':scope > references > reference').forEach((tag) => {
-          references[tag.id] = tag.innerHTML;
-          config.references[tag.id] = applyDeserializedXMLTransforms(deserialize(tag.innerHTML), {
-            arrays,
-            renameTable
-          });
-        });
-        // Replacing references
-        xml.querySelectorAll('configuration > reference').forEach((tag) => {
-          tag.outerHTML = references[tag.id];
-        });
-        // Make sure any plugin reference has a valid site id to import the plugin from
-        xml.querySelectorAll('plugin').forEach((tag) => {
-          const siteAttr = tag.getAttribute('site');
-          if (siteAttr === '{site}' || siteAttr === null) {
-            tag.setAttribute('site', site);
-          }
-        });
-        const toolbar = xml.querySelector('[id="craftercms.components.PreviewToolbar"] > configuration');
-        if (toolbar) {
-          const leftSection = toolbar.querySelector('leftSection > widgets');
-          if (leftSection) {
-            leftSection.querySelectorAll('widget').forEach((e, index) => e.setAttribute('uiKey', String(index)));
-            config.preview.toolbar.leftSection = applyDeserializedXMLTransforms(deserialize(leftSection), {
-              arrays,
-              renameTable
-            });
-          }
-          const middleSection = toolbar.querySelector('middleSection > widgets');
-          if (middleSection) {
-            middleSection.querySelectorAll('widget').forEach((e, index) => e.setAttribute('uiKey', String(index)));
-            config.preview.toolbar.middleSection = applyDeserializedXMLTransforms(deserialize(middleSection), {
-              arrays,
-              renameTable
-            });
-          }
-          const rightSection = toolbar.querySelector('rightSection > widgets');
-          if (rightSection) {
-            rightSection.querySelectorAll('widget').forEach((e, index) => e.setAttribute('uiKey', String(index)));
-            config.preview.toolbar.rightSection = applyDeserializedXMLTransforms(deserialize(rightSection), {
-              arrays,
-              renameTable
-            });
-          }
-        }
-        const toolsPanelPages = xml.querySelector('[id="craftercms.components.ToolsPanel"] > configuration > widgets');
-        if (toolsPanelPages) {
-          // When rendering widgets dynamically and changing pages on the tools panel, if there are duplicate react key
-          // props across pages, react may no swap the components correctly, incurring in unexpected behaviours.
-          // We need a unique key for each widget.
-          toolsPanelPages.querySelectorAll('widget').forEach((e, index) => e.setAttribute('uiKey', String(index)));
-          const lookupTables = ['fields'];
-          config.preview.toolsPanel = applyDeserializedXMLTransforms(deserialize(toolsPanelPages), {
-            arrays,
-            lookupTables,
-            renameTable
-          });
-        }
-        const launcher = xml.querySelector('[id="craftercms.components.Launcher"] > configuration');
-        if (launcher) {
-          launcher.querySelectorAll('widget').forEach((e, index) => e.setAttribute('uiKey', String(index)));
-          config.launcher = applyDeserializedXMLTransforms(deserialize(launcher), {
-            arrays,
-            renameTable
-          }).configuration;
-        }
-        const pageBuilderPanel = xml.querySelector(
-          '[id="craftercms.components.PageBuilderPanel"] > configuration > widgets'
-        );
-        if (pageBuilderPanel) {
-          const lookupTables = ['fields'];
-          pageBuilderPanel.querySelectorAll('widget').forEach((e, index) => {
-            e.setAttribute('uiKey', String(index));
-            if (e.getAttribute('id') === 'craftercms.components.ToolsPanelPageButton') {
-              e.querySelector(':scope > configuration')?.setAttribute('target', 'pageBuilderPanel');
-            }
-          });
-          config.preview.pageBuilderPanel = applyDeserializedXMLTransforms(deserialize(pageBuilderPanel), {
-            arrays,
-            renameTable,
-            lookupTables
-          });
-        }
-        const dashboard = xml.querySelector('[id="craftercms.components.Dashboard"] > configuration');
-        if (dashboard) {
-          dashboard.querySelectorAll('widget').forEach((e, index) => e.setAttribute('uiKey', String(index)));
-          config.dashboard = applyDeserializedXMLTransforms(deserialize(dashboard), {
-            arrays,
-            renameTable
-          }).configuration;
-        }
-        return config;
-      } else {
-        return null;
-      }
-    })
-  );
+export function fetchSiteUiConfig(site: string): Observable<string> {
+  // const arrays = ['widgets', 'roles', 'excludes', 'devices', 'values', 'siteCardMenuLinks', 'tools'];
+  // const renameTable = { permittedRoles: 'roles' };
+  // Reading references
+  // const references = {};
+  // xml.querySelectorAll(':scope > references > reference').forEach((tag) => {
+  //   references[tag.id] = tag.innerHTML;
+  //   config.references[tag.id] = applyDeserializedXMLTransforms(deserialize(tag.innerHTML), {
+  //     arrays,
+  //     renameTable
+  //   });
+  // });
+  // // Replacing references
+  // // TODO: referencesExpansion will go in xmlPreprocessor
+  // xml.querySelectorAll('configuration > reference').forEach((tag) => {
+  //   tag.outerHTML = references[tag.id];
+  // });
+  // xml.querySelectorAll('plugin').forEach((tag) => {
+  //   const siteAttr = tag.getAttribute('site');
+  //   if (siteAttr === '{site}' || siteAttr === null) {
+  //     tag.setAttribute('site', site);
+  //   }
+  // });
+
+  return fetchConfigurationXML(site, '/ui.xml', 'studio');
 }
 
 const legacyToNextMenuIconMap = {
