+++ ui/app/src/utils/hooks/useSiteUIConfig.ts
@@ -19,8 +19,7 @@ import { useActiveSiteId } from './useActiveSiteId';
 import { useDispatch } from 'react-redux';
 import { useSelection } from './useSelection';
 import { useEffect } from 'react';
-import { fetchSiteUiConfig, fetchSiteUiConfig2Complete } from '../../state/actions/configuration';
-import { fetchSiteUiConfig2 } from '../../services/configuration';
+import { fetchSiteUiConfig } from '../../state/actions/configuration';
 
 export function useSiteUIConfig(): GlobalState['uiConfig'] {
   const site = useActiveSiteId();
@@ -29,9 +28,6 @@ export function useSiteUIConfig(): GlobalState['uiConfig'] {
   useEffect(() => {
     if (config.currentSite !== site && !config.isFetching) {
       dispatch(fetchSiteUiConfig({ site }));
-
-      // TODO: this is just for testing, needs to be an action and handled as currently is with fetchSiteUiConfig
-      fetchSiteUiConfig2(site).subscribe((config) => dispatch(fetchSiteUiConfig2Complete(config)));
     }
   }, [dispatch, site, config.isFetching, config.currentSite]);
   return config;
